requirements.txt
pandas
openpyxl
snowflake-connector-python
python-dotenv